Handle abstraction of quotesHelper functions
